Reuse calculations across dimensions.
The prior algorithm generated a separate set of samples for each number
of dimensions, making it O(p * d^2 * r). Because we don't need the
samples to be independent for each number of dimensions tested, however,
we can generate a new sum of coordinates for the N-dimensional case
dimension by merely adding one more term to the (N-1)-dimensional case,
allowing us to generate the table in O(p * d * r) time instead.